# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: School Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students (filtered by school for librarians)
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /students/{id}/history:
    get:
      operationId: getStudentBorrowHistory
      tags: [students]
      summary: Get borrow history for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student borrow history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IssueRecord"

  # ─── ISSUES ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Student:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        name:
          type: string
        rollNumber:
          type: string
        class:
          type: string
        section:
          type: string
        contactPhone:
          type: string
      required: [id, schoolId, name, rollNumber, class, section, contactPhone]
    CreateStudentBody:
      type: object
      properties:
        schoolId:
          type: string
        name:
          type: string
        rollNumber:
          type: string
        class:
          type: string
        section:
          type: string
        contactPhone:
          type: string
      required: [schoolId, name, rollNumber, class, section, contactPhone]
    IssueRecord:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        bookId:
          type: string
        studentId:
          type: string
        issuedBy:
          type: string
        issueDate:
          type: string
        dueDate:
          type: string
        returnDate:
          type: string
          nullable: true
        fineAmount:
          type: number
        status:
          type: string
          enum: [issued, returned, overdue]
        bookTitle:
          type: string
        bookAuthor:
          type: string
        studentName:
          type: string
        studentRollNumber:
          type: string
        studentClass:
          type: string
        studentSection:
          type: string
      required: [id, schoolId, bookId, studentId, issuedBy, issueDate, dueDate, fineAmount, status, bookTitle, bookAuthor, studentName, studentRollNumber, studentClass, studentSection]
...